Divide connector examples into Pages (#659)
* update sandbox example to useES connector

* fixed type issue with wasSearched boolean state

* Elasticsearch Connector POC

* allow apiKey to be passed in

* add debugger configuration

* update elasticsearch connector

* split out ResponseState types

* add tests + refactor elasticsearch connector

* remove server

* update searchkit-sdk dep

* fix type issues

* fix data model errors in tests

* readme documentation

* remove proxy configuration + update field config

* update readme

* WIP

* typo on privileges.


* accidental copy


* Update packages/search-ui-elasticsearch-connector/README.md


* Update packages/search-ui-elasticsearch-connector/README.md


* remove npm-run-all where not needed; clean up scripts

* refactor around the handler methods to be a configuration object

* re-enable location facet + update configuration code to use geodistance facet

* Fix Pagination issue + tests for range facets

* split out examples into pages, inline config in page
