+++ examples/sandbox/src/pages/elasticsearch/index.js
@@ -1,9 +1,9 @@
 import React from "react";
 import "@elastic/eui/dist/eui_theme_light.css";
 
-import AppSearchAPIConnector from "@elastic/search-ui-app-search-connector";
-import SiteSearchAPIConnector from "@elastic/search-ui-site-search-connector";
 import ElasticSearchAPIConnector from "@elastic/search-ui-elasticsearch-connector";
+import moment from "moment";
+
 import {
   ErrorBoundary,
   Facet,
@@ -23,22 +23,135 @@ import {
   SingleSelectFacet
 } from "@elastic/react-search-ui-views";
 import "@elastic/react-search-ui-views/lib/styles/styles.css";
-import {
-  config as ElasticSearchConfig,
-  fields as ElasticsearchFields
-} from "./configurations/Elasticsearch";
-import {
-  config as EntSearchConfig,
-  fields as EntSearchFields
-} from "./configurations/EntSearch";
-import "./App.css";
 
-const sourceMode = process.env.REACT_APP_SOURCE;
+const connector = new ElasticSearchAPIConnector(
+  {
+    host: process.env.REACT_ELASTICSEARCH_HOST || "http://localhost:9200",
+    index: process.env.REACT_ELASTICSEARCH_INDEX || "us_parks"
+  },
+  {
+    queryFields: ["title", "description", "states"]
+  }
+);
 
-const isElasticsearchSource =
-  sourceMode && sourceMode.indexOf("ELASTICSEARCH") !== -1;
-
-const fields = isElasticsearchSource ? ElasticsearchFields : EntSearchFields;
+const config = {
+  debug: true,
+  alwaysSearchOnInitialLoad: true,
+  apiConnector: connector,
+  hasA11yNotifications: true,
+  searchQuery: {
+    result_fields: {
+      visitors: { raw: {} },
+      world_heritage_site: { raw: {} },
+      location: { raw: {} },
+      acres: { raw: {} },
+      square_km: { raw: {} },
+      title: {
+        snippet: {
+          size: 100,
+          fallback: true
+        }
+      },
+      nps_link: { raw: {} },
+      states: { raw: {} },
+      date_established: { raw: {} },
+      image_url: { raw: {} },
+      description: {
+        snippet: {
+          size: 100,
+          fallback: true
+        }
+      }
+    },
+    disjunctiveFacets: ["acres", "states.keyword", "date_established", "location"],
+    facets: {
+      "world_heritage_site.keyword": { type: "value" },
+      "states.keyword": { type: "value", size: 30 },
+      acres: {
+        type: "range",
+        ranges: [
+          { from: -1, name: "Any" },
+          { from: 0, to: 1000, name: "Small" },
+          { from: 1001, to: 100000, name: "Medium" },
+          { from: 100001, name: "Large" }
+        ]
+      },
+      location: {
+        // San Francisco. In the future, make this the user's current position
+        center: "37.7749, -122.4194",
+        type: "range",
+        unit: "mi",
+        ranges: [
+          { from: 0, to: 100, name: "Nearby" },
+          { from: 100, to: 500, name: "A longer drive" },
+          { from: 500, name: "Perhaps fly?" }
+        ]
+      },
+      date_established: {
+        type: "range",
+        ranges: [
+          {
+            from: moment()
+              .subtract(50, "years")
+              .toISOString(),
+            name: "Within the last 50 years"
+          },
+          {
+            from: moment()
+              .subtract(100, "years")
+              .toISOString(),
+            to: moment()
+              .subtract(50, "years")
+              .toISOString(),
+            name: "50 - 100 years ago"
+          },
+          {
+            to: moment()
+              .subtract(100, "years")
+              .toISOString(),
+            name: "More than 100 years ago"
+          }
+        ]
+      },
+      visitors: {
+        type: "range",
+        ranges: [
+          { from: 0, to: 10000, name: "0 - 10000" },
+          { from: 10001, to: 100000, name: "10001 - 100000" },
+          { from: 100001, to: 500000, name: "100001 - 500000" },
+          { from: 500001, to: 1000000, name: "500001 - 1000000" },
+          { from: 1000001, to: 5000000, name: "1000001 - 5000000" },
+          { from: 5000001, to: 10000000, name: "5000001 - 10000000" },
+          { from: 10000001, name: "10000001+" }
+        ]
+      }
+    }
+  },
+  autocompleteQuery: {
+    results: {
+      resultsPerPage: 5,
+      result_fields: {
+        title: {
+          snippet: {
+            size: 100,
+            fallback: true
+          }
+        },
+        nps_link: {
+          raw: {}
+        }
+      }
+    },
+    suggestions: {
+      types: {
+        documents: {
+          fields: ["states.keyword"]
+        }
+      },
+      size: 4
+    }
+  }
+};
 
 const SORT_OPTIONS = [
   {
@@ -49,7 +162,7 @@ const SORT_OPTIONS = [
     name: "Title",
     value: [
       {
-        field: fields.title,
+        field: "title.keyword",
         direction: "asc"
       }
     ]
@@ -58,7 +171,7 @@ const SORT_OPTIONS = [
     name: "State",
     value: [
       {
-        field: fields.states,
+        field: "states.keyword",
         direction: "asc"
       }
     ]
@@ -67,11 +180,11 @@ const SORT_OPTIONS = [
     name: "State -> Title",
     value: [
       {
-        field: fields.states,
+        field: "states.keyword",
         direction: "asc"
       },
       {
-        field: fields.title,
+        field: "title.keyword",
         direction: "asc"
       }
     ]
@@ -80,74 +193,30 @@ const SORT_OPTIONS = [
     name: "Heritage Site -> State -> Title",
     value: [
       {
-        field: fields.world_heritage_site,
+        field: "world_heritage_site.keyword",
         direction: "asc"
       },
       {
-        field: fields.states,
+        field: "states.keyword",
         direction: "asc"
       },
       {
-        field: fields.title,
+        field: "title.keyword",
         direction: "asc"
       }
     ]
   }
 ];
 
-let connector = null;
-let queryConfig = null;
-
-if (sourceMode === "ELASTICSEARCH_CONNECTOR") {
-  connector = new ElasticSearchAPIConnector(
-    {
-      host: process.env.REACT_ELASTICSEARCH_HOST || "http://localhost:9200",
-      index: process.env.REACT_ELASTICSEARCH_INDEX || "us_parks"
-    },
-    {
-      queryFields: ["title", "description", "states"]
-    }
-  );
-  queryConfig = ElasticSearchConfig;
-} else if (process.env.REACT_APP_SOURCE === "SITE_SEARCH") {
-  connector = new SiteSearchAPIConnector({
-    engineKey:
-      process.env.REACT_SITE_SEARCH_ENGINE_KEY || "Z43R5U3HiDsDgpKawZkA",
-    documentType: process.env.REACT_SITE_SEARCH_ENGINE_NAME || "national-parks"
-  });
-  queryConfig = EntSearchConfig;
-} else {
-  connector = new AppSearchAPIConnector({
-    searchKey:
-      process.env.REACT_APP_SEARCH_KEY || "search-371auk61r2bwqtdzocdgutmg",
-    engineName:
-      process.env.REACT_APP_SEARCH_ENGINE_NAME || "search-ui-examples",
-    hostIdentifier:
-      process.env.REACT_APP_SEARCH_HOST_IDENTIFIER || "host-2376rb",
-    endpointBase: process.env.REACT_APP_SEARCH_ENDPOINT_BASE || ""
-  });
-  queryConfig = EntSearchConfig;
-}
-
-const config = {
-  debug: true,
-  alwaysSearchOnInitialLoad: true,
-  ...queryConfig,
-  apiConnector: connector,
-  hasA11yNotifications: true
-};
-
 export default function App() {
   return (
     <SearchProvider config={config}>
       <WithSearch
-        mapContextToProps={({ wasSearched, authorizeUrl, isLoggedIn }) => ({
-          wasSearched,
-          authorizeUrl,
-          isLoggedIn
+        mapContextToProps={({ wasSearched }) => ({
+          wasSearched
         })}
       >
-        {({ wasSearched, authorizeUrl, isLoggedIn }) => {
+        {({ wasSearched }) => {
           return (
             <div className="App">
               <ErrorBoundary>
@@ -173,13 +242,13 @@ export default function App() {
                         <Sorting label={"Sort by"} sortOptions={SORT_OPTIONS} />
                       )}
                       <Facet
-                        field={fields.states}
+                        field="states.keyword"
                         label="States"
                         filterType="any"
                         isFilterable={true}
                       />
                       <Facet
-                        field={fields.world_heritage_site}
+                        field="world_heritage_site.keyword"
                         label="World Heritage Site"
                         view={BooleanFacet}
                       />
